unstable: detection area enlarged, one entity = one block
First attempt: very strange behaviour. Too many strange bugs. This commit is only here for later research.
I should first focus on the default mechanic working properly, instead of trying to fix the placement-API.

This branch will be happening inside of witchtrading as this mechanic is already featured there through the sad_orchid. Here, I'll try to fix the placement-API (which the sad_orchid makes use off) in a proper, non-hacky and reliable way. If this branch is done, it will be merged to master and new features will be added in other datapacks using this API.