voiddimension: intorucing voiddimension!
The voiddimension has nothing in it, its just there. Which means that the mechanics are working and content can now be implemented. There are still missing whole mechanics, but I will count those as further content to be implemented yet.
Further information about this can be read in the r2 project https://github.com/Metroite/datapacks/projects/2

improvements:
*Shinigami Power would create smoke, even while just flying around. It is now score-based, won't happen anymore.
*Batwings can now also be eaten for saturation, not just healing
*reworked the sapphire_glass dupe-protection, it is now less friendly to dupes
*removed worseoldpvp out of the hurtfulpack, this will have to be installed seperately - it doesn't fit into the pack