voiddimension: its now possible to enter and exit the dimension
to enter, you have to make a so called grave by placing two soul_sand blocks next to each other and planting wither_roses on both of them. Now place a skull in head height (2 blocks) looking at you. The skull has to look straight and has to be placed straight (not diagonal). While standing on one of the wither_roses look at the skull for 16 seconds. After some effects and damage due to the wither_rose (take golden_apples with you), you will be teleported into the void dimension.

to exit, you have to lower your health to half a heart through the wither effect, which you can take with you as the wither suspicious stew made with wither_roses. As soon as the player reaches half a heart and should actually die, the player will be teleported back to the overworld and won't die at all.

improvements:
*limitedlife: buffed the Batwing Suit
*voiddimension: the slow_falling effect now holds until the player reaches the ground
*voiddimension: players in the void dimension will be in adventuremode (I'm thinking about a concept for the dimension)