voiddimension: increased performance and center can be changed now
set $shift_x$ and $shift_z$ in vd.coordinates a different value to change the center of the void. $travel$ can also be changed which is calculating how much a block in the void equals to blocks in the overworld. 1 block in the void = 16 block in the overworld in the default settings.

performance has probably been increases as i discovered that there is a /execute if/unless score <target> MATCHES <value> syntax. lol.