torchout: implemented custom durability mechanic
*now also runs slightly more often to increase reliability

furthermore, as Mojang decided to remove the ability to modify player data completely, I have decided to enable the so called "bukkit" support concerning the arrow mechanics and the old/alternative method for worseoldpvp by default. too bad...
If you are using Minecraft 1.14 without Bukkit, you should turn them off by running following commands first for worseoldpvp and second for the arrow mechanics:
/scoreboard players set $oldmethod$ wop.extend 0
/scoreboard players set $bukkit$ hp.storeboard 0

(btw. every change in this month was always exclusive to the hurfulpack for now, my next goal is to bring the changes to their standalone packs)